requirements.txt
fastapi
sqlalchemy
python-dotenv
cryptography
requests
uvicorn